requirements.txt
async-timeout==4.0.3
click==7.1.2
fastapi==0.63.0
fastapi-cache2==0.2.1
fastapi-jwt-auth==0.5.0
greenlet==3.0.3
h11==0.14.0
mysql-connector-python==8.4.0
pendulum==3.0.0
pydantic==1.8.2
PyJWT==1.7.1
python-dateutil==2.9.0.post0
redis==5.0.5
six==1.16.0
SQLAlchemy==1.4.11
starlette==0.13.6
time-machine==2.14.1
typing_extensions==4.12.2
tzdata==2024.1
uvicorn==0.13.4
websockets==9.1
